None of this works
Google deprecated the login method used by Net::Google::Spreadsheets

This looks like it could be re-worked, but for the sake of time,
I'm going to go with CSV import/export for the time being.

Given how we intend to invoice customers, it may be useful to have
direct API access to Google Drive in the future.