Added new fields for proper invoicing
debit -> how much this account should be debited
for this invoice period.  This is the column to use
to add up all the invoices to see what the user has been billed.
Individual invoice amounts can be higher if the family falls
behind or smaller if they are ahead, but this amount must be the
amount owed for this invoice period in order for invoices and payments
to add up.

adjust -> Notification to the mail-merging person to possibly adjust
this invoice due to an outstanding balance or a change in the family's
tuition or payment plan