A database change and a parsing problem
I'm not 100% on where it happened, but everyone in
Advanced Band ended up with just Band as their class.
Put in a note to investigate.

Changed the names of the days of the week to numbers
so that the sort order is correct.  Made a note of that here
as well so this utility will be updated to store days of the
week numerically for next year.