Changed default nobill value from null to ''
The nobill field is used to determine if an account
should be charged and or sent invoices.  It's also useful
to translate it's contents into something more readable
for statements and things.  Having nulls in here made the
lookups out of the prettify table fail, causing the rest of
the row to not appear in query output.  I've not found a root
cause of this yet, but it obviously has something to do with
the way SQLite handles nulls.