Add a column to track student status
Should be null if student is active in class, anything else
if they are not.

Students can add and withdraw from classes at any time.
Use this column to track that status.

Students that are inactive will still be counted in total
number of students. However if email lists are re-generated, they
will respect this status.