This is the single worst commit of the whole project
I knew ahead of time that I could run into problems
with decimal dollar amounts and floating point math.
I figured since the number of transactions per account
was small, I wouldn't run into any issues.  This is untrue.

The system should be redesigned to have an integer number of
cents, or 10ths or cents, or 100ths of cents, something like
that.  Then when a dollar value is needed, the query or
querying code will need to do the math.  All calculations should
be done on this integer of cents and the conversion to
dollars should only be done for display purposes.

I was warned, and I still thought I knew better.

https://blog.plover.com/prog/Moonpig.html