Fully abandon broken Google Sheets interface
Output will be in XLSX or CSV, depending on formatting needs.

There may be some other modules.  I dunno if there's a good
Google Spreadsheet one, but we may be able to use Drive::Simple
or something like it to mirror a local dir and get data into
Google Sheets that way.