Removing obsolete utilities
download-emails just got a list of emails from the signup sheet
since the parser wasn't done yet.  We're better off just parsing
the signups now that it works and making the lists we want from
there.

The family roster isn't useful and was requested to be dropped for
this year.

The provisional roster is made from the signup sheet directly. Since we
can't get at Google Sheets from Perl currently, it's of no utility.
And we only really used it because the parser wasn't totally done on
time last year.