Set nobill correctly
We use an empty string to represent the default billing status.
We use nulls everywhere else, but this was convenient for the
making the statement utility.  We may want to revisit this as
this is the only place we use empty strings and not nulls in the
db and that might be a bad thing overall.