Transition to new views
This helps consolidate the logic for determining which families and
people are in the active session and which students are active.

By creating these views with all of the constraints in one place,
we can easily change the definition of these things and we can
use that logic appropriately anywhere we need it in the programs.

Billing people are one idea of an "active" person.  This definition
should be used when trying to apply payments or look up families that
may owe.  Any logic needed in a financial sense should use billing_*.

Ed_student is used to see who's an "active" student from a educational
standpoint.  When determining things like which parents should appear
on which class list or if a family has any active students, this view
should be used as it is sensitive to kids withdrawing and such.

Basically once a family is signed up, they are a billing_family and will
stay one for the duration of their session.  Students can come and go.