Respect validation status and the dates on entries
Anything that isn't a debit needs to be validated, since there
is a chance they were entered erroneously.  All of them are done
by hand.

And we should respect dates.  This would allow us to pre-enter debits
or credits and not have it effect the family's balance until such
time that it should take effect.