Don't notify on credits that don't pay off the account
We don't want send an email saying "Thank you for your payment"
when we issue a credit.  The family doesn't see it as a payment,
so they would be confused.  But we do want to send a "Thanks for
paying your tuition in full" if the account is paid off, just
so they know they are covered.