Added "macro" type to make-email-lists
Macros are bits of SQL (in this case, where clauses)
queried out of the DB so that I don't have to re-jigger
the list generator.  I could have made an object with
prepare_cached, execute, and fetch methods, but doing it
with an actual statement object seemed like less work, if a
bit hackey.

It might be a good idea to change the table name to
macro_lists or something, and then have a type there
that can be queried out.  Then I wouldn't have to make
more tables for a different macro.

I tried to use a SQL placeholder ("?") to hold the place for
the clauses, but things bound to placeholders are quoted, so
it broke the queries. I ended up with the %%QUERY%% marker
and simple text substitution because, well, it was easy.
I mean, if you are going to do SQL injection, you gotta do some
SQL injection, amirite?